% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{add}
\alias{add}
\title{Adds a state block or a measurement equation to a given state space model}
\usage{
add(model, item)
}
\arguments{
\item{model}{A state space model}

\item{item}{A state block or a measurement equation}
}
\description{
Adds a state block or a measurement equation to a given state space model
}
\examples{
model<-model()
llt<-locallineartrend("llt")
add(model,llt)
}
