% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{ar}
\alias{ar}
\alias{ar2}
\title{Autoregressive model}
\usage{
ar(
  name,
  ar,
  fixedar = FALSE,
  variance = 0.01,
  fixedvariance = FALSE,
  nlags = 0,
  zeroinit = FALSE
)

ar2(
  name,
  ar,
  fixedar = FALSE,
  variance = 0.01,
  fixedvariance = FALSE,
  nlags = 0,
  nfcasts = 0
)
}
\arguments{
\item{ar}{vector of the AR coefficients (See @details
Additional details...).}

\item{fixedar}{boolean that triggers the estimation of the AR coefficients (\code{FALSE})
or fixed it (\code{TRUE}) to a pre-specified value set by the parameter \code{ar}.}

\item{variance}{the variance (\eqn{\sigma^2_{ar}}).}

\item{fixedvariance}{boolean that triggers the estimation of the variance (\code{FALSE})
or fixed it (\code{TRUE}) to a pre-specified value set by the parameter \code{variance}.}

\item{nlags}{integer specifying how many lags of the state variable are needed}

\item{zeroinit}{boolean determining the initial condition for the state variable,
which is equal to zero if \code{zeroinit = TRUE}.
The default (\code{zeroinit = FAKSE}) triggers the an initialization based on the
unconditional mean and variance of the AR(p) process.}

\item{nfcasts}{integer specifying how many forecasts of the state variable are needed}
}
\value{
An element of type "JD3_SsfStateBlock" (wrapper around the corresponding Java object))
}
\description{
Functions to create an autoregressive model (\code{ar}) or a
modified autoregressive model (\code{ar2})
}
\details{
The AR process is defined by
\deqn{y_t = \phi_1 y_{t-1} +  \phi_2 y_{t-2} + \dots + \phi_p y_{t-p} + \epsilon_t}
The stability of the auto-regressive polynomial is not checked.
}
\examples{
b_ar<-ar("my_ar", c(.8,-.3,.2), variance=1)
block_p0(b_ar)
}
