% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{arima}
\alias{arima}
\title{Autoregressive Integrated Moving Average (ARIMA) Model}
\usage{
arima(name, ar, diff, ma, var = 1, fixed = FALSE)
}
\description{
Autoregressive Integrated Moving Average (ARIMA) Model
}
