% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{arma}
\alias{arma}
\title{Autoregressive Moving Average (ARMA) Model}
\usage{
arma(name, ar, fixedar = FALSE, ma, fixedma = FALSE, var = 1, fixedvar = FALSE)
}
\description{
Autoregressive Moving Average (ARMA) Model
}
