% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{bsm_model}
\alias{bsm_model}
\title{Title}
\usage{
bsm_model(
  period,
  noise.var = 1,
  level.var = 1,
  slope.var = NULL,
  seasonal.var = 1,
  seasonal.model = c("HarrisonStevens", "Trigonometric", "Dummy", "Crude", "Fixed"),
  cycle.var = NULL,
  cycle.dumpingfactor = 0.9,
  cycle.length = 6
)
}
\description{
Title
}
