% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{components_pos}
\alias{components_pos}
\title{Position of the components}
\usage{
components_pos(model)
}
\arguments{
\item{model}{Estimated model}
}
\value{
The first position of the different blocks in the state array
}
\description{
Position of the components
}
\examples{
model<-model()
ll<-locallevel("ll")
seas<-seasonal("seas", 12, "Crude")
n<-noise("n")
add(model,ll)
add(model,seas)
add(model,n)
emodel<-estimate(model, rjd3toolkit::Retail$BookStores)
print(components_pos(emodel))
}
