% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{compute}
\alias{compute}
\title{Computes a model, for given parameters}
\usage{
compute(model, data, parameters, marginal = FALSE, concentrated = TRUE)
}
\arguments{
\item{model}{the model}

\item{data}{a matrix containing the data (one time series per column, time series dimension on the rows)}

\item{parameters}{Parameters of the model}

\item{marginal}{logical value used to specify whether the marginal likelihood definition is used (TRUE) or
not (FALSE) during the optimization. The marginal likelihood is recommended when there is at least one variable that loads
on a non-stationary latent variable and the loading coefficient needs to be estimated.}

\item{concentrated}{logical value used to specify whether the likelihood is concentrated (TRUE) or not (FALSE) during the optimization}
}
\value{
An object of the class "JD3_SsfModelEstimation"
}
\description{
Computes a model, for given parameters
}
