% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{cycle}
\alias{cycle}
\title{Title}
\usage{
cycle(
  name,
  factor = 0.9,
  period = 60,
  fixed = FALSE,
  variance = 0.01,
  fixedvariance = FALSE
)
}
\arguments{
\item{name}{Name of the block}

\item{variance}{Variance of the innovations}

\item{fixedvariance}{Indicates if the variance if fixed}
}
\description{
Title
}
