% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_states.R
\name{.airline}
\alias{.airline}
\title{Creates the state space form of an airline model;}
\usage{
.airline(period = 12, theta = -0.6, btheta = -0.8)
}
\arguments{
\item{period}{Period of the model.}

\item{theta}{Regular moving average parameter.}

\item{btheta}{Seasonal moving average parameter.}
}
\value{
The raw state space form of the Airline model.
}
\description{
Creates the state space form of an airline model;
}
\examples{
ssf<-.airline(24)
.dk_likelihood(ssf, rjd3toolkit::ABS$X0.2.09.10.M)
}
