% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.akf_likelihood}
\alias{.akf_likelihood}
\title{Title}
\usage{
.akf_likelihood(
  ssf,
  data,
  qtype = c("NORMAL", "PARTIAL_TRIANGULARIZATION", "FULL_TRIANGULARIZATION", "QR"),
  collapsing = TRUE,
  rescalingFactor = TRUE
)
}
\arguments{
\item{qtype}{}
}
\description{
Title
}
