% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{.bsm2spec}
\alias{.bsm2spec}
\title{Title}
\usage{
.bsm2spec(jbsm, fixed = FALSE, fixedCycle = TRUE)
}
\arguments{
\item{fixedCycle}{}
}
\description{
Title
}
\examples{
bsm<-bsm_model(12)
jbsm<-.r2jd_bsm(bsm)
jspec<-.bsm2spec(jbsm)
}
