% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{.bsm2ucm}
\alias{.bsm2ucm}
\title{Title}
\usage{
.bsm2ucm(jbsm, normalized = TRUE)
}
\arguments{
\item{normalized}{}
}
\description{
Title
}
\examples{
bsm<-bsm_model(12)
jbsm<-.r2jd_bsm(bsm)
jucm<-.bsm2ucm(jbsm)
}
