% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_algorithms.R
\name{.dk_likelihood}
\alias{.dk_likelihood}
\title{Computes the diffuse likelihood by means of the diffuse Kalman filter (Durbin-Koopman).}
\usage{
.dk_likelihood(ssf, data, sqr = TRUE, rescalingFactor = TRUE)
}
\arguments{
\item{ssf}{A state space form.}

\item{data}{Arrays of data.}

\item{sqr}{Square root diffuse initialization, if any}

\item{rescalingFactor}{True if the innovation covariance matrix is defined up to a scaling factor.}
}
\value{
The diffuse likelihood
}
\description{
Computes the diffuse likelihood by means of the diffuse Kalman filter (Durbin-Koopman).
}
\examples{
sarima<-.sarima(12, NULL, 1,  -.6, NULL, 1, -.5)
ssf<-.ssf(sarima, .loading(0))
.dk_likelihood(ssf, rjd3toolkit::ABS$X0.2.09.10.M)
}
