% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_loadings.R
\name{.loading_composite}
\alias{.loading_composite}
\title{Creates a composite loading}
\usage{
.loading_composite(dims, cmps)
}
\arguments{
\item{dims}{The dimensions of the corresponding state blocks}

\item{cmps}{The loadings}
}
\description{
Creates a composite loading
}
\examples{
l<-.loading_composite(c(1,2,3), list(.loading(0), .loading(0), .loading(0)))
.ssf_Z(l, 6, 0)
}
