% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_states.R
\name{.noise}
\alias{.noise}
\title{Creates a white noise.}
\usage{
.noise(var = 1)
}
\arguments{
\item{var}{Variance of the noise.}
}
\value{
A wrapper around the java object (class JD3_RawStateBlock).
}
\description{
Creates a white noise.
}
\examples{
sb<-.noise(.01)
.ssf_T(sb, 0)
}
