% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.sarima}
\alias{.sarima}
\title{Creates an ARMA state block}
\usage{
.sarima(period, phi, d = 1, theta, bphi, bd = 1, btheta)
}
\arguments{
\item{period}{Period of the seasonality}

\item{phi}{Regular stationary auto-regressive polynomial. True signs, without the constant}

\item{d}{Regular differencing order}

\item{theta}{Regular moving average polynomial. True signs, without the constant}

\item{bphi}{Seasonal stationary auto-regressive polynomial. True signs, without the constant}

\item{bd}{Seasonal differencing order}

\item{btheta}{Seasonal moving average polynomial. True signs, without the constant}
}
\value{
A raw java state block
}
\description{
Creates an ARMA state block
}
\examples{
sb<-.sarima(12, c(0.5, -.2, .1), 1, -.8, NULL, 1, -.6)
.ssf_P0(sb)
}
