% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.ssf_smooth}
\alias{.ssf_smooth}
\title{Title}
\usage{
.ssf_smooth(
  ssf,
  data,
  all = TRUE,
  qtype = c("NORMAL", "PARTIAL_TRIANGULARIZATION", "FULL_TRIANGULARIZATION", "QR")
)
}
\arguments{
\item{qtype}{}
}
\description{
Title
}
