% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.ssf_V}
\alias{.ssf_V}
\title{Gets the covariance of the innovations in the transition equation.}
\usage{
.ssf_V(x, pos = 0)
}
\arguments{
\item{x}{A state block or a state space form.}

\item{pos}{The 0-based position (use 0 for time invariant models)}
}
\value{
The requested matrix.
}
\description{
Gets the covariance of the innovations in the transition equation.
}
\examples{
sarima<-.sarma(12, .9, -.6, .9, -.5)
.ssf_V(sarima)
}
