% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.ssf_Z}
\alias{.ssf_Z}
\title{Gets the loading vector. It should have the same length as the corresponding state block.}
\usage{
.ssf_Z(x, n, pos)
}
\arguments{
\item{x}{Loading}

\item{n}{the Length of the corresponding state block (on which the loading will be applied)}

\item{pos}{0-based position of the filter}
}
\value{
the loading vector
}
\description{
Gets the loading vector. It should have the same length as the corresponding state block.
}
