% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.state_dim}
\alias{.state_dim}
\title{Retrieves the dimension of a state block}
\usage{
.state_dim(x)
}
\arguments{
\item{x}{A state block}
}
\value{
The length of the state block
}
\description{
Retrieves the dimension of a state block
}
\examples{
s<-.seasonal(12)
.state_dim(s)
}
