% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.state_of}
\alias{.state_of}
\title{Gets the state of the state space form}
\usage{
.state_of(ssf)
}
\arguments{
\item{ssf}{A state space form}
}
\value{
A state block
}
\description{
Gets the state of the state space form
}
