% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_sts_option}
\alias{get_sts_option}
\title{Set an option for sts}
\usage{
get_sts_option(name)
}
\arguments{
\item{name}{Name of the option}
}
\value{
The requested option or NULL if it doesn't exist
}
\description{
Set an option for sts
}
\examples{
sts_option("test", "DUMMY")
get_sts_option("test")
}
