% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{loadings}
\alias{loadings}
\title{Give all the loadings for a given variable}
\usage{
loadings(object, var = 1)
}
\arguments{
\item{object}{Estimated model}

\item{var}{loadings for the given variable (or equation). Useless in the case of univariate models}
}
\value{
A matrix with the requested loadings. The number of rows corresponds to the number of observations.
}
\description{
Give all the loadings for a given variable
}
\examples{
model<-model()
llt<-locallineartrend("llt")
seas<-seasonal("seas", 12, "Crude")
n<-noise("n")
add(model,llt)
add(model,seas)
add(model,n)
y<-rjd3toolkit::Retail$BookStores
emodel<-estimate(model, y)
print(loadings(emodel))
}
