% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{noise}
\alias{noise}
\title{Noise state block}
\usage{
noise(name, variance = 0.01, fixed = FALSE)
}
\arguments{
\item{name}{Name of the block}

\item{variance}{Variance of the noise}

\item{fixed}{Indicates if the variance is fixed}
}
\description{
Noise state block
}
\examples{
n<-noise("n", 1)
block_t(n)
}
