% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{reg}
\alias{reg}
\title{Time Varying Regressors}
\usage{
reg(name, x, var = NULL, fixed = FALSE)
}
\arguments{
\item{x}{matrix containing the regressors}

\item{fixed}{}
}
\description{
Time Varying Regressors
}
