% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{reg_td}
\alias{reg_td}
\title{Title}
\usage{
reg_td(
  name,
  period,
  start,
  length,
  groups = c(1, 2, 3, 4, 5, 6, 0),
  contrast = TRUE,
  variance = 1,
  fixed = FALSE
)
}
\arguments{
\item{name}{}

\item{period}{}

\item{start}{}

\item{length}{}

\item{groups}{}

\item{contrast}{}

\item{variance}{}

\item{fixed}{}
}
\description{
Title
}
