% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{seasonal}
\alias{seasonal}
\title{Seasonal state block}
\usage{
seasonal(
  name,
  period,
  type = c("Trigonometric", "Crude", "HarrisonStevens", "Dummy"),
  variance = 0.01,
  fixed = FALSE
)
}
\arguments{
\item{name}{Name of the block}

\item{period}{Period of the seasonality}

\item{type}{Type of the seasonal component}

\item{variance}{Innovation variance}

\item{fixed}{Indicates if the variance is fixed}
}
\description{
Seasonal state block
}
\examples{
seas1<-seasonal("seas1", 12, "HarrisonStevens", variance=1)
print(block_v(seas1))
seas2<-seasonal("seas2", 12, "Trigonometric", variance=1)
print(block_v(seas2))
}
