% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{smoothed_components}
\alias{smoothed_components}
\title{Retrieves the components of the model (univariate case) or the
components corresponding to a given equation (multivariate case)}
\usage{
smoothed_components(model, equation = 1, fast = FALSE)
}
\arguments{
\item{model}{Estimated state space model}

\item{equation}{Equation containing the components}

\item{fast}{if true, only the components are computed. Otherwise, their stdev
are also computed (not returned but available for future use).}
}
\value{
A matrix with the components
}
\description{
Retrieves the components of the model (univariate case) or the
components corresponding to a given equation (multivariate case)
}
\examples{
model<-model()
llt<-locallineartrend("llt")
seas<-seasonal("seas", 12, "Crude")
n<-noise("n")
add(model,llt)
add(model,seas)
add(model,n)
y<-rjd3toolkit::Retail$BookStores
emodel<-estimate(model, y)
scmp<-smoothed_components(emodel)
high<-cbind(scmp[,1], y-scmp[,2])
low<-scmp[,c(2,3)]
matplot(high, type='l')
matplot(low, type='l')
}
