% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{splines_generic}
\alias{splines_generic}
\title{Title}
\usage{
splines_generic(
  name,
  period,
  knots,
  order = 4,
  start = 1,
  variance = 1,
  fixed = FALSE
)
}
\arguments{
\item{fixed}{}
}
\description{
Title
}
