% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{splines_regular}
\alias{splines_regular}
\title{Title}
\usage{
splines_regular(
  name,
  period,
  nknots = 0,
  knots = NULL,
  start = 1,
  variance = 1,
  fixed = FALSE
)
}
\arguments{
\item{fixed}{}
}
\description{
Title
}
