% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stsoutliers.R
\name{sts.outliers}
\alias{sts.outliers}
\title{Title}
\usage{
sts.outliers(
  y,
  period = NA,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  ao = T,
  ls = T,
  so = F,
  cv = 0,
  tcv = 0,
  estimation.forward = c("Score", "Point", "Full"),
  estimation.backward = c("Point", "Score", "Full")
)
}
\arguments{
\item{y}{input time series.}

\item{period}{annual frequency.}

\item{X}{Regression variables (same length as y) or NULL}

\item{X.td}{Specification for trading days clustering.
Contains thr group id fo Mondays... Sundays
(for instance (1,1,1,1,1,0,0) for week days or (1,1,1,1,1,2,0) for week.saturdays/Sundays variables).
Contrasts are used. Can be NULL}

\item{level}{-1 = no level, 0 = fixed level, 1 = sotchastic level}

\item{slope}{}

\item{noise}{}

\item{seasonal}{Seasonal model}

\item{ao, ls, so}{boolean indicating if additive outliers (\code{ao}), level shift (\code{ls}) and seasonal
outliers (\code{so}) should be detected.}

\item{cv}{}

\item{tcv}{}

\item{estimation.forward}{}

\item{estimation.backward}{}
}
\description{
Title
}
\examples{
 x<-rjd3toolkit::retail$BookStores
 sts.outliers(x)
}
