% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{sts}
\alias{sts}
\title{Title}
\usage{
sts(
  y,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  cycle = -1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  diffuse.regs = TRUE,
  tol = 1e-09
)
}
\arguments{
\item{y}{input time series.}

\item{X}{Regression variables (same length as y) or NULL}

\item{X.td}{Groups of days for trading days regressors. The length of the array must be 7.
It indicates to what group each week day belongs.
The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0. The other groups are identified by 1, 2,... n (<= 6).
For instance, usual trading days are defined by \code{c(1,2,3,4,5,6,0)}, week days by \code{c(1,1,1,1,1,0,0)}, etc...}

\item{level}{-1 = no level, 0 = fixed level, 1 = sotchastic level}

\item{seasonal}{Seasonal model}

\item{tol}{}
}
\description{
Title
}
\examples{
 x<-rjd3toolkit::Retail$BookStores
 sts(x)
}
