% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{sts_forecast}
\alias{sts_forecast}
\title{Forecast with STS model}
\usage{
sts_forecast(y, model = c("none", "td2", "td3", "td7", "full"), nf = 12)
}
\arguments{
\item{y}{Series}

\item{model}{Model for calendar effects
\itemize{
\item{td2: leap year + week days (week-end derived)}
\item{td3: leap year + week days + saturdays (sundays derived)}
\item{td7: leap year + all days (sundays derived)}
\item{full: td3 + easter effect}
\item{none: no calendar effect}
}}

\item{nf}{number of forecasts}
}
\description{
Forecast with STS model
}
\examples{
fcasts<-sts_forecast(rjd3toolkit::ABS$X0.2.09.10.M)
}
