% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sts_option}
\alias{sts_option}
\title{Set an option for sts}
\usage{
sts_option(name, obj)
}
\arguments{
\item{name}{Name of the option}

\item{obj}{Option}
}
\description{
Set an option for sts
}
\examples{
sts_option("test", "DUMMY")
}
