% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{sts_raw}
\alias{sts_raw}
\title{Title}
\usage{
sts_raw(
  y,
  period = NA,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  cycle = -1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  diffuse.regs = TRUE,
  tol = 1e-09
)
}
\description{
Title
}
