% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{var_ar}
\alias{var_ar}
\title{Title}
\usage{
var_ar(
  name,
  ar,
  fixedar = FALSE,
  stderr,
  scale = 1,
  fixed = FALSE,
  zeroinit = FALSE
)
}
\arguments{
\item{name}{Name of the block}
}
\description{
Title
}
