% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{var_locallevel}
\alias{var_locallevel}
\title{Title}
\usage{
var_locallevel(name, std, scale = 1, fixed = FALSE, initial = NaN)
}
\arguments{
\item{name}{}

\item{std}{}

\item{scale}{}

\item{fixed}{}

\item{initial}{}
}
\description{
Title
}
