% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{var_locallineartrend}
\alias{var_locallineartrend}
\title{Title}
\usage{
var_locallineartrend(
  name,
  lstd,
  sstd = NULL,
  levelScale = 1,
  slopeScale = 1,
  fixedLevelScale = FALSE,
  fixedSlopeScale = FALSE
)
}
\description{
Title
}
