% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{var_seasonal}
\alias{var_seasonal}
\title{Title}
\usage{
var_seasonal(
  name,
  period,
  type = c("Trigonometric", "Crude", "HarrisonStevens", "Dummy"),
  std,
  scale = 1,
  fixed = FALSE
)
}
\description{
Title
}
