% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{varlocallevel}
\alias{varlocallevel}
\title{Title}
\usage{
varlocallevel(name, std, scale = 1, fixed = FALSE, initial = NaN)
}
\arguments{
\item{name}{}

\item{std}{}

\item{scale}{}

\item{fixed}{}

\item{initial}{}
}
\description{
Title
}
