% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{varlocallineartrend}
\alias{varlocallineartrend}
\title{Title}
\usage{
varlocallineartrend(
  name,
  lstd,
  sstd = NULL,
  levelScale = 1,
  slopeScale = 1,
  fixedLevelScale = FALSE,
  fixedSlopeScale = FALSE
)
}
\arguments{
\item{name}{}

\item{lstd}{}

\item{sstd}{}

\item{levelScale}{}

\item{slopeScale}{}

\item{fixedLevelScale}{}

\item{fixedSlopeScale}{}
}
\description{
Title
}
