% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{varreg}
\alias{varreg}
\title{Title}
\usage{
varreg(name, x, stderr, scale = 1, fixed = F)
}
\arguments{
\item{name}{}

\item{x}{Regression variable. Numerics}

\item{stderr}{Standard error of the innovations of the coefficient (1 in extrapolation)}

\item{scale}{Scaling factor}

\item{fixed}{Fixed scaling factor}
}
\value{

}
\description{
Title
}
\examples{
 x<-rjd3toolkit::retail$BookStores
 std<-rep(1, length(x))
 std[c(20, 50, 150)]<-5
 v<-varreg("vx", x, std, 0.1)
}
