% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{varseasonal}
\alias{varseasonal}
\title{Title}
\usage{
varseasonal(
  name,
  period,
  type = c("Trigonometric", "Crude", "HarrisonStevens", "Dummy"),
  std,
  scale = 1,
  fixed = FALSE
)
}
\arguments{
\item{name}{}

\item{period}{}

\item{type}{}

\item{std}{}

\item{scale}{}

\item{fixed}{}
}
\value{

}
\description{
Title
}
