% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{add.equation}
\alias{add.equation}
\title{Title}
\usage{
add.equation(equation, item, coeff = 1, fixed = TRUE, loading = NULL)
}
\arguments{
\item{equation}{}

\item{item}{}

\item{coeff}{}

\item{fixed}{}

\item{loading}{}
}
\description{
Title
}
