% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{add_equation}
\alias{add_equation}
\title{Add Latent variable}
\usage{
add_equation(equation, item, coeff = 1, fixed = TRUE, loading = NULL)
}
\arguments{
\item{equation}{the equation}

\item{coeff}{the value of the coefficient associated to the block of latent variables defined by \code{item}.}

\item{fixed}{logical that triggers estimation of coeff (FALSE) or fixes it (TRUE) to a pre-specified value}

\item{loading}{}
}
\description{
Add Latent variable
}
