% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{arma}
\alias{arma}
\title{Autoregressive Moving Average (ARMA) Model}
\usage{
arma(name, ar, fixedar = FALSE, ma, fixedma = FALSE, var = 1, fixedvar = FALSE)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{fixedar}{}

\item{ma}{}

\item{fixedma}{}

\item{var}{}

\item{fixedvar}{}
}
\description{
Autoregressive Moving Average (ARMA) Model
}
