% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_algorithms.R
\name{.akf_likelihood}
\alias{.akf_likelihood}
\title{Computes the diffuse likelihood by means of the augmented Kalman filter.}
\usage{
.akf_likelihood(
  ssf,
  data,
  qtype = c("NORMAL", "PARTIAL_TRIANGULARIZATION", "FULL_TRIANGULARIZATION", "QR"),
  collapsing = TRUE,
  rescalingFactor = TRUE
)
}
\arguments{
\item{ssf}{A state space form.}

\item{data}{Arrays of data.}

\item{qtype}{Type of the initialization in the augmented Kalman filter}

\item{rescalingFactor}{True if the innovation covariance matrix is defined up to a scaling factor.}
}
\value{
The diffuse likelihood
}
\description{
Computes the diffuse likelihood by means of the augmented Kalman filter.
}
\examples{
sarima<-.sarima(12, NULL, 1,  -.6, NULL, 1, -.5)
ssf<-.ssf(sarima, .loading(0))
.akf_likelihood(ssf, rjd3toolkit::ABS$X0.2.09.10.M)
}
