% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf.R
\name{.noise}
\alias{.noise}
\title{Creates a white noise}
\usage{
.noise(var = 1)
}
\arguments{
\item{var}{Variance of the noise}
}
\value{
A raw java state block
}
\description{
Creates a white noise
}
\examples{
sb<-.noise(.01)
.ssf_T(sb, 0)
}
