% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_states.R
\name{.sarma}
\alias{.sarma}
\title{Creates an ARMA state block (representation I)}
\usage{
.sarma(period, phi, theta, bphi, btheta)
}
\arguments{
\item{period}{Period of the seasonality}

\item{phi}{Regular stationary auto-regressive polynomial. True signs, without the constant}

\item{theta}{Regular moving average polynomial. True signs, without the constant}

\item{bphi}{Seasonal stationary auto-regressive polynomial. True signs, without the constant}

\item{btheta}{Seasonal moving average polynomial. True signs, without the constant}
}
\value{
A raw java state block
}
\description{
Creates an ARMA state block (representation I)
}
\examples{
sb<-.sarma(12, c(0.5, -.2, .1), -.8, NULL, -.6)
.ssf_P0(sb)
}
