% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_states.R
\name{.seasonal}
\alias{.seasonal}
\title{Creates a seasonal component, corresponding to a multivariate random walk,
with an aggregation constraint to 0 and various covariances for the innovations of the transition equation.}
\usage{
.seasonal(
  period,
  type = c("Trigonometric", "Crude", "HarrisonStevens", "Dummy"),
  var = 1
)
}
\arguments{
\item{period}{Period of the seasonality}

\item{type}{Type of the innovations of the transition equation}

\item{var}{Variance of the innovations}
}
\value{
A wrapper around the java object (class JD3_RawStateBlock)
}
\description{
Creates a seasonal component, corresponding to a multivariate random walk,
with an aggregation constraint to 0 and various covariances for the innovations of the transition equation.
}
\examples{
sb<-.seasonal(4, "HarrisonStevens", .01)
.ssf_V(sb, 0)
}
