% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_rawssf_algorithms.R
\name{.ssf_smooth}
\alias{.ssf_smooth}
\title{Computes smoothed states by means of the augmented Kalman filter in the case
of diffuse initialization}
\usage{
.ssf_smooth(
  ssf,
  data,
  all = TRUE,
  qtype = c("NORMAL", "PARTIAL_TRIANGULARIZATION", "FULL_TRIANGULARIZATION", "QR")
)
}
\arguments{
\item{ssf}{A state space form.}

\item{data}{Arrays of data.}

\item{all}{True if the covariances of the smoothed states are compted.}

\item{qtype}{Type of the initialization in the augmented Kalman filter}
}
\value{
A matrix with the smoothed states and - if requested - their standard deviations.
}
\description{
Computes smoothed states by means of the augmented Kalman filter in the case
of diffuse initialization
}
