% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{estimate}
\alias{estimate}
\title{Title}
\usage{
estimate(
  model,
  data,
  marginal = F,
  concentrated = T,
  initialization = c("Augmented_Robust", "Diffuse", "SqrtDiffuse", "Augmented",
    "Augmented_NoCollapsing"),
  optimizer = c("LevenbergMarquardt", "MinPack", "BFGS", "LBFGS"),
  precision = 1e-15,
  initialParameters = NULL
)
}
\arguments{
\item{model}{}

\item{data}{}

\item{marginal}{}

\item{concentrated}{}

\item{initialization}{}

\item{optimizer}{}

\item{precision}{}

\item{initialParameters}{}
}
\value{

}
\description{
Title
}
