% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{loading}
\alias{loading}
\title{Title}
\usage{
loading(pos = NULL, weights = NULL)

loading(pos = NULL, weights = NULL)
}
\arguments{
\item{pos}{}

\item{weights}{}

\item{object}{}

\item{obs}{}
}
\description{
Title

Title
}
