% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{loading}
\alias{loading}
\title{Title}
\usage{
loading(pos = NULL, weights = NULL)

loading(pos = NULL, weights = NULL)
}
\arguments{
\item{pos}{defines the position of each one of the elements of the block of states defined.
\code{NULL} indicates by default the first state included in the block (pos=0)}

\item{weights}{defines the weights associated to each one of the state variables included in the block.}

\item{object}{}

\item{obs}{}
}
\description{
Title

Title
}
