% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{periodic}
\alias{periodic}
\title{Title}
\usage{
periodic(name, period, harmonics, variance = 0.01, fixedvariance = FALSE)
}
\arguments{
\item{name}{}

\item{period}{}

\item{harmonics}{}

\item{variance}{}

\item{fixedvariance}{}
}
\description{
Title
}
