% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{signal}
\alias{signal}
\title{Title}
\usage{
signal(object, obs = 1, pos = NULL, loading = NULL, stdev = F)
}
\arguments{
\item{object}{}

\item{obs}{}

\item{pos}{}

\item{loading}{}

\item{stdev}{}
}
\description{
Title
}
