% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{splines_regular}
\alias{splines_regular}
\title{Title}
\usage{
splines_regular(
  name,
  period,
  nnodes = 0,
  nodes = NULL,
  start = 1,
  variance = 1,
  fixed = FALSE
)
}
\arguments{
\item{name}{}

\item{period}{}

\item{nnodes}{}

\item{nodes}{}

\item{start}{}

\item{variance}{}

\item{fixed}{}
}
\description{
Title
}
