% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{sts.raw}
\alias{sts.raw}
\title{Title}
\usage{
sts.raw(
  y,
  period = NA,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  cycle = -1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  diffuse.regs = T,
  tol = 1e-09
)
}
\arguments{
\item{y}{input time series.}

\item{X}{Regression variables (same length as y) or NULL}

\item{X.td}{Specification for trading days clustering.
Contains thr group id fo Mondays... Sundays
(for instance (1,1,1,1,1,0,0) for week days or (1,1,1,1,1,2,0) for week.saturdays/Sundays variables).
Contrasts are used. Can be NULL}

\item{level}{-1 = no level, 0 = fixed level, 1 = sotchastic level}

\item{slope}{}

\item{cycle}{}

\item{noise}{}

\item{seasonal}{Seasonal model}

\item{diffuse.regs}{}

\item{tol}{}
}
\description{
Title
}
