% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{var_reg}
\alias{var_reg}
\title{Time Varying Regressor}
\usage{
var_reg(name, x, stderr, scale = 1, fixed = FALSE)
}
\arguments{
\item{name}{}

\item{x}{Regression variable. Numerics}

\item{stderr}{Standard error of the innovations of the coefficient (1 in extrapolation)}

\item{scale}{Scaling factor}

\item{fixed}{Fixed scaling factor}
}
\description{
Time Varying Regressor
}
\examples{
 x<-rjd3toolkit::retail$BookStores
 std<-rep(1, length(x))
 std[c(20, 50, 150)]<-5
 v<-var_reg("vx", x, std, 0.1)
}
